inventor (2.1.5-10-2) unstable; urgency=low
  * debian/control: New package inventor-doc for documentation.
    Program ivman used to generate manpages only built on
    certain architectures.  Closes: #240927.

  * doc/GNUmakefile: Generate manpages only if requested.

  * doc/man/ivman: Fixes to make 'g++ -Wall' quiet.

 -- Steve M. Robbins <smr@debian.org>  Tue,  1 Jun 2004 22:24:10 -0400

inventor (2.1.5-10-1) unstable; urgency=low

  * New upstream version.  Note that the .orig.tgz file is modified
  from the upstream version: I had to remove all CVS directories.

  * Replaced all usage of "acroread" by "xpdf".  Closes: #227794.

  * debian/control: -dev package depends on lesstif-dev.  Closes: #237699.

 -- Steve M. Robbins <smr@debian.org>  Fri, 28 May 2004 22:19:43 -0400

+++ doc/man/ivman/Man.c++
@@ -806,7 +806,7 @@ Method::parse()
 
 #define NUM_OP_NAMES (sizeof(opNames) / sizeof(opNames[0]))
 
-		int op;
+		size_t op;
 		for (op = 0; op < NUM_OP_NAMES; op++) {
 		    if (name == opNames[op][0]) {
 			name = opNames[op][1];
